# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/evidence": {
      "get": {
        "operationId": "EvidenceController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `guest`, `superuser`, `compliance_officer`, `project_manager`, `hr_manager`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get all evidence",
        "tags": [
          "Evidence"
        ]
      },
      "post": {
        "operationId": "EvidenceController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `compliance_officer`, `project_manager`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created."
          }
        },
        "summary": "Submit new evidence",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}": {
      "get": {
        "operationId": "EvidenceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `guest`, `superuser`, `compliance_officer`, `project_manager`, `hr_manager`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Get evidence by ID",
        "tags": [
          "Evidence"
        ]
      },
      "patch": {
        "operationId": "EvidenceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `compliance_officer`, `project_manager`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEvidenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Update evidence status/content",
        "tags": [
          "Evidence"
        ]
      },
      "delete": {
        "operationId": "EvidenceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `compliance_officer`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation."
          }
        },
        "summary": "Delete evidence",
        "tags": [
          "Evidence"
        ]
      }
    },
    "/evidence/{id}/upload": {
      "post": {
        "operationId": "EvidenceController_uploadFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "description": "**Role-Based Access Control**\nAllowed roles: `superuser`, `compliance_officer`, `project_manager`, `team_leader`, `team_member`",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "x-user-id",
            "description": "String user id for the acting user (required on task/subtask/escalation mutations)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-company-id",
            "description": "UUID of the company context (required for tenant routes)",
            "schema": {}
          },
          {
            "in": "header",
            "required": false,
            "name": "x-platform-admin-id",
            "description": "UUID of the platform admin user (required for platform routes)",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Evidence file (PDF, DOCX, PNG, etc.)"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Optional notes about this upload"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded, Attachment created, and linked to evidence record."
          }
        },
        "summary": "Upload a file to an evidence record (real Multer upload)",
        "tags": [
          "Evidence"
        ]
      }
    },
...
  },
  "info": {
    "title": "OfficeSync API",
    "description": "Backend API for the OfficeSync HR and PM Dashboard",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Role-Based-Access": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "Enter your role to unlock the endpoints (e.g., platform_admin, superuser, hr_manager, team_leader, team_member)"
      }
    },
    "schemas": {
...
      "CreateEvidenceDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "comp-uuid-1"
          },
          "user_id": {
            "type": "string",
            "example": "user-uuid-9"
          },
          "task_id": {
            "type": "string",
            "example": "task-uuid-104"
          },
          "violation_id": {
            "type": "string",
            "example": "violation-uuid-2"
          },
          "title": {
            "type": "string",
            "example": "Server patch test results"
          },
          "evidence_type": {
            "type": "string",
            "example": "Document"
          },
          "file_url": {
            "type": "string",
            "example": "https://storage.officesync.in/evidence/report.pdf"
          },
          "notes": {
            "type": "string",
            "example": "Partial test results from staging."
          }
        },
        "required": [
          "companyId",
          "user_id",
          "title",
          "file_url"
        ]
      },
      "UpdateEvidenceDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "example": "comp-uuid-1"
          },
          "user_id": {
            "type": "string",
            "example": "user-uuid-9"
          },
          "task_id": {
            "type": "string",
            "example": "task-uuid-104"
          },
          "violation_id": {
            "type": "string",
            "example": "violation-uuid-2"
          },
          "title": {
            "type": "string",
            "example": "Server patch test results"
          },
          "evidence_type": {
            "type": "string",
            "example": "Document"
          },
          "file_url": {
            "type": "string",
            "example": "https://storage.officesync.in/evidence/report.pdf"
          },
          "notes": {
            "type": "string",
            "example": "Partial test results from staging."
          },
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Under_Review",
              "Approved",
              "Rejected"
            ],
            "example": "Approved"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "Role-Based-Access": []
    }
  ]
}